refactor(web): extract Select (Alpha) typeahead into a useTypeahead hook
